requirements.txt
fastapi
python-dotenv
pytest
uvicorn
sendgrid
openai-swarm